Move timeslot components into subdirectory
