+++ src/components/Session/index.js
@@ -3,6 +3,9 @@ import withAuthentication from "./withAuthentication";
 import withAuthorization from "./withAuthorization";
 
 const isAdmin = (authUser) => authUser && !!authUser.roles.admin;
+const isRecruitmentTeam = (authUser) =>
+  authUser && !!authUser.roles.recruitmentteam;
+const isApplicant = (authUser) => authUser && !!authUser.roles.applicant;
 const isLoggedIn = (authUser) => !!authUser;
 
 export {
@@ -10,5 +13,7 @@ export {
   withAuthentication,
   withAuthorization,
   isAdmin,
+  isRecruitmentTeam,
+  isApplicant,
   isLoggedIn,
 };
