+++ src/components/Timeslots/ScheduleColumn.js
