+++ src/components/Timeslots/index.js
@@ -1,0 +1,4 @@
+import InterviewerView from "./InterviewerView";
+import ApplicantView from "./ApplicantView";
+
+export { InterviewerView, ApplicantView };
